+++ irs-edc-client/src/test/java/org/eclipse/tractusx/irs/edc/client/SubmodelRetryerTest.java
@@ -84,7 +84,7 @@ class SubmodelExponentialRetryTest {
         final EdcSubmodelClient client = new EdcSubmodelClientImpl(config, negotiationService, dataPlaneClient,
                 endpointDataReferenceStorage,
                 pollingService, retryRegistry, catalogFacade, endpointDataReferenceCacheService);
-        testee = new EdcSubmodelFacade(client);
+        testee = new EdcSubmodelFacade(client, config);
     }
 
     @Test