+++ irs-registry-client/src/test/java/org/eclipse/tractusx/irs/registryclient/DefaultConfigurationTest.java
@@ -35,6 +35,7 @@ import java.util.List;
 import java.util.concurrent.ExecutionException;
 
 import org.eclipse.edc.spi.types.domain.edr.EndpointDataReference;
+import org.eclipse.tractusx.irs.edc.client.EdcConfiguration;
 import org.eclipse.tractusx.irs.edc.client.EdcSubmodelClient;
 import org.eclipse.tractusx.irs.edc.client.EdcSubmodelFacade;
 import org.eclipse.tractusx.irs.edc.client.exceptions.EdcClientException;
@@ -43,7 +44,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.web.client.RestTemplate;
 
 class DefaultConfigurationTest {
-
+    private final EdcConfiguration edcConfiguration = new EdcConfiguration();
     private final DefaultConfiguration testee = new DefaultConfiguration();
     private final String descriptorTemplate = "descriptor/{aasIdentifier}";
     private final String shellLookupTemplate = "shell?{assetIds}";
@@ -70,7 +71,7 @@ class DefaultConfigurationTest {
     @Test
     void edcSubmodelFacade() {
         final EdcSubmodelClient facadeMock = mock(EdcSubmodelClient.class);
-        final EdcSubmodelFacade edcSubmodelFacade = testee.edcSubmodelFacade(facadeMock);
+        final EdcSubmodelFacade edcSubmodelFacade = testee.edcSubmodelFacade(facadeMock,edcConfiguration);
 
         assertThat(edcSubmodelFacade).isNotNull();
     }