+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed submodel request path by introducing configuration property `irs-edc-client.submodel.submodel-suffix` which will
   be appended to the href URL.
+- Fixed validation of json-schemas - IRS is creating tombstone instead collecting Submodel payload, when it not passes validation of schema #522
 
 ## [4.5.2] - 2024-02-22
 ## Changed